bugs arreglados

+++ core/src/main/java/com/traceability/core/domain/physicalasset/payloads/AssetDepletedPayload.java
@@ -2,10 +2,12 @@ package com.traceability.core.domain.physicalasset.payloads;
 
 import com.traceability.core.domain.event.DomainEventPayload;
 
+import java.math.BigDecimal;
+
 /**
  * Payload for ASSET_DEPLETED event.
  * Ref: ADR-005
  */
 public record AssetDepletedPayload(
-    long previousQuantity
+    BigDecimal previousQuantity
 ) implements DomainEventPayload {}
