+++ core/src/main/java/com/traceability/core/domain/physicalasset/payloads/AssetRegisteredPayload.java
@@ -2,6 +2,8 @@ package com.traceability.core.domain.physicalasset.payloads;
 
 import com.traceability.core.domain.event.DomainEventPayload;
 
+import java.math.BigDecimal;
+
 /**
  * Payload for ASSET_REGISTERED event.
  * Ref: ADR-002, ADR-014
@@ -9,7 +11,7 @@ import com.traceability.core.domain.event.DomainEventPayload;
 public record AssetRegisteredPayload(
     String assetId,
     String assetType,
-    long quantity,
+    BigDecimal quantity,
     String unitOfMeasure,
     String currentLocation,
     String custodianRef,
