+++ core/src/main/java/com/traceability/core/domain/physicalasset/payloads/AssetSplitCompensatedPayload.java
@@ -2,11 +2,13 @@ package com.traceability.core.domain.physicalasset.payloads;
 
 import com.traceability.core.domain.event.DomainEventPayload;
 
+import java.math.BigDecimal;
+
 /**
  * Payload for ASSET_SPLIT_COMPENSATED event.
  * Ref: ADR-008, ADR-009
  */
 public record AssetSplitCompensatedPayload(
     String childAssetId,
-    long reintegratedQuantity
+    BigDecimal reintegratedQuantity
 ) implements DomainEventPayload {}
