+++ core/src/main/java/com/traceability/core/domain/physicalasset/payloads/AssetSplitPayload.java
@@ -2,16 +2,18 @@ package com.traceability.core.domain.physicalasset.payloads;
 
 import com.traceability.core.domain.event.DomainEventPayload;
 
+import java.math.BigDecimal;
+
 /**
  * Payload for ASSET_SPLIT event.
  * Ref: ADR-005, ADR-008
  */
 public record AssetSplitPayload(
     String childAssetId,
-    long extractedQuantity,
+    BigDecimal extractedQuantity,
     String unitOfMeasure,
-    long parentQuantityBefore,
-    long parentQuantityAfter,
+    BigDecimal parentQuantityBefore,
+    BigDecimal parentQuantityAfter,
     String statusBeforeSplit,
     String childLocation,
     String childCustodianRef,
